requirements.txt
Django==5.2.6
channels==4.1.0
asgiref==3.8.1
gunicorn==21.2.0
uvicorn[standard]==0.24.0
django-apscheduler==0.7.0
APScheduler==3.11.0
django-mongodb-backend==5.2.1
pymongo==4.15.1
pandas==2.3.2
pytz==2025.2
python-decouple==3.8
